+ finished jest config + worked out file structure for testing + example component updated to adhere to current repo guidelines
