--- components/example/example.test.tsx
@@ -1,15 +1,0 @@
-import React from 'react'
-import { render, screen } from '@testing-library/react'
-import Example from ".";
-
-describe('Example', () => {
-    it('renders a heading', () => {
-        render(<Example />)
-
-        const heading = screen.getByRole('heading', {
-            name: /Welcome to next\.js!/i,
-        })
-
-        expect(heading).toBeInTheDocument()
-    })
-})
