--- components/example/index.tsx
@@ -1,11 +1,0 @@
-import styles from './example.module.scss';
-
-const Example = () => {
-    return (
-        <h3 className={styles.redHeading}>
-            Welcome to <a href="https://nextjs.org">Next.js!</a>
-        </h3>
-    )
-};
-
-export default Example;
