+++ src/__tests__/utils/index.ts
@@ -1,0 +1,5 @@
+import React from "react";
+
+it('is a default test', () => {
+    expect(1 + 1).toBe(2);
+});
