+++ src/components/App/Example/index.tsx
@@ -1,0 +1,11 @@
+import styles from './Example.module.scss';
+
+const Example = () => {
+  return (
+    <h3 className={styles.redHeading}>
+      Welcome to <a href="https://nextjs.org">Next.js!</a>
+    </h3>
+  );
+};
+
+export default Example;
